+++ app/assets/javascripts/components/story/StoryHistoryLocation.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import StoryActionButton from './StoryActionButton'
 
 const StoryHistoryLocation = ({ id, url }) =>
   <div className="col-xs-12 form-group input-group input-group-sm">
@@ -13,26 +14,31 @@ const StoryHistoryLocation = ({ id, url }) =>
       <button
         className="btn btn-default btn-clipboard"
         data-clipboard-target={`#story-link-${id}`}
-        type="button"
+        type="button" 
+        title={I18n.t('story.events.copy_url')}
       >
-        <img src="/clippy.svg" alt="Copy to clipboard" className="clipboard-icon" />
+        <img src="/clippy.svg" alt={I18n.t('story.events.copy_url')} className="clipboard-icon" />
       </button>
 
       <button
         className="btn btn-default btn-clipboard-id btn-clipboard"
         data-clipboard-text={`#${id}`}
         type="button"
+        title={I18n.t('story.events.copy_id')}
       >
         ID
       </button>
 
-      <button className="btn btn-default toggle-history">
-        <i className="mi md-18">history</i>
-      </button>
-
-      <button className="btn btn-default clone-story" title={I18n.t('story.events.clone')}>
-        <i className="mi md-18">content_copy</i>
-      </button>
+   <StoryActionButton
+      className={'toggle-history'}
+      title={I18n.t('story.events.view_history')}
+      iconName={'history'}
+    />
+    <StoryActionButton 
+      className={'clone-story'}
+      title={I18n.t('story.events.clone')}
+      iconName={'content_copy'}
+    />
     </span>
   </div>